pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "webverse"
version = "1.0"
description = "Local GUI lab runner for web hacking labs using Docker Compose v2."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Leighlin Ramsay"}]
dependencies = [
    "PyQt5>=5.15.9",
    "PyYAML>=6.0",
    "packaging>=24.0",
    "requests==2.32.3",
    "httpx==0.27.2",
    "flask==3.0.3",
    "PyJWT==2.9.0",
    "passlib==1.7.4",
    "bcrypt==4.2.0",
    "python-multipart==0.0.12",
    "starlette==0.41.3",
    "jinja2==3.1.4",
    "itsdangerous==2.2.0",
]

[project.scripts]
webverse = "webverse.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
webverse = [
    "labs/**",
    "gui/icons/**",
    "gui/**/*.png",
    "gui/**/*.svg",
    "gui/**/*.ico",
]
